Extract Account model in support file
