+++ Gemfile
@@ -3,6 +3,7 @@ source 'https://rubygems.org'
 gemspec
 
 gem "activerecord", "> 4.0.6", "< 4.2"
+gem "mysql2", '~> 0.3.0', :platforms => :ruby
 gem "protected_attributes"
 
 eval(File.read('gemfiles/common.rb'))
