+++ test/support/acts_like_an_integer.rb
@@ -1,0 +1,5 @@
+class ActsLikeAnInteger
+  def to_i
+    123
+  end
+end
